+++ tests/test_lazy_loading.py
@@ -13,9 +13,12 @@ def test_requirements_single_marker_lazy_load():
     assert roi_requires_single_marker_load(35000000, 55)
     assert not roi_requires_single_marker_load(array, 1)
     assert not roi_requires_single_marker_load(np.zeros((1000, 1000)), 1000)
+    assert roi_requires_single_marker_load(np.zeros((1000, 1000)), 2000)
     assert not roi_requires_single_marker_load(10000, 100)
 
     assert roi_requires_single_marker_load(2e10, 1)
+    assert not roi_requires_single_marker_load((5200*3000), 54)
+    assert roi_requires_single_marker_load((8800*9500), 47)
 
 def test_lazy_loading_single_marker_mcd(get_current_dir):
     parser = FileParser([os.path.join(get_current_dir, "query.mcd")])